Minor changes: -Added a "Close Button" to each form -Added a Application close button -Hide the ControlMenu of the Main Form -Added DialogResult to ask again if the Customer should be deleted -Changed behaviour of add customer button: now all TextBoxes will be cleared and you can add a new Customer directly. You close the form now via "Close button"

+++ Kundenkartei/MainForm.cs
@@ -196,9 +196,18 @@ namespace Kundenkartei
             if (metroListView1.CheckedItems.Count == 1)
             {
                 string kundenNr = metroListView1.CheckedItems[0].SubItems[1].Text;
-                SqliteDataAccess.DeleteKunde(kundenNr);
-                metroListView1.Items.Clear();
-                FillCustomerList();
+                DialogResult dialogResult = MessageBox.Show("Kunde wirklich löschen?", "Kunde löschen", MessageBoxButtons.YesNo);
+                if (dialogResult == DialogResult.Yes)
+                {                    
+                    SqliteDataAccess.DeleteKunde(kundenNr);
+                    metroListView1.Items.Clear();
+                    FillCustomerList();
+                }
+                else if (dialogResult == DialogResult.No)
+                {
+                    //do something else
+                }
+                
             }
             else
             {
@@ -372,6 +381,11 @@ namespace Kundenkartei
             {
                 MessageBox.Show("Bitte einen Kunden auswählen");
             }
+        }
+
+        private void metroButton7_Click(object sender, EventArgs e)
+        {
+            Environment.Exit(0);
         }
     }
 }
